refactor: Prefer <img /> over <Image /> for cards
Due to how many cards (~250+) there are in each magic set (some with
multiple faces) and there being ~150 sets we offer access to, we quickly
hit the image optimization limit offered by Vercel: 1000 for hobby tier.

https://vercel.com/dashboard/usage#Image%20Optimization-Count

Instead of leveraging <Image /> for cards, we'll just use <img /> for
now and postpone this problem for later.