fix: API not available at build time (see more)
The reason the last several commits didn't fix the build
is because the API is never available at build time.

This comment led me to this realization: https://github.com/vercel/next.js/discussions/13316#discussioncomment-6265173

> During the build, there is no server running. All server components are prerendered so they can be served during runtime/request time. So the reference to localhost during buildtime isn't going to work unless you have a dedicated server running on that port.
>
> if you want to use the /api route use it with a real host that can serve your data to prerender pages.
> /api routes that point to localhost should be used only with client side during runtime/request time ('use client'). At this point you have a server running at localhost that can serve your data.
>
> I think its an issue with your mental model of hoe nextjs works- not an issue with nextjs.

So in effect, we need to directly talk to the Scryfall API for
any server components to work.