fix(types): Cypress and Jest clashing types
https://docs.cypress.io/guides/tooling/typescript-support#Clashing-types-with-Jest

TypeScript was not compiling due to Cypress and Jest having overlapping globals (e.g. `describe`). The best solution I found was to provide a separate `tsconfig.json` specifically for cypress.

There was another approach that involved using the package [`local-cypress`](https://www.npmjs.com/package/local-cypress), but I was not a fan of the side-effects.